server/requirements.txt
anyio==3.6.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
